yarn format and removed legacy/scenario.ts
