--- _common/gameModel/libs/ClientScript/game/legacy/scenario.ts
@@ -1,17 +1,0 @@
-import { EventPayload, ScriptedEvent } from '../common/events/eventTypes';
-
-export function reviveScriptedEvent(
-  emitter: {
-    emitterCharacterId: string;
-    emitterPlayerId: string;
-  },
-  targetId: string,
-  scripted: ScriptedEvent
-): EventPayload {
-  const pe: EventPayload = {
-    ...emitter,
-    ...scripted.payload,
-    targetId: targetId,
-  };
-  return pe;
-}
