+++ basic_scenario/gameModel/gamemodel.json
@@ -6354,15 +6354,6 @@
   }, {
     "@class" : "GameModelContent",
     "content" : "",
-    "contentKey" : "game/legacy/scenario",
-    "contentType" : "application/typescript",
-    "libraryType" : "ClientScript",
-    "refId" : "GameModelContent:#69:OCwcyC",
-    "version" : null,
-    "visibility" : "PRIVATE"
-  }, {
-    "@class" : "GameModelContent",
-    "content" : "",
     "contentKey" : "game/legacy/the_world",
     "contentType" : "application/typescript",
     "libraryType" : "ClientScript",
